requirements.txt
click==8.1.7
aiohttp==3.9.3
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.5.14
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.0
faiss-cpu==1.8.0
fastapi==0.100.1
fastapi-socketio==0.0.10
frozenlist==1.4.1
fsspec==2024.2.0
googleapis-common-protos==1.62.0
grpcio==1.62.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.21.4
idna==3.6
importlib-metadata==6.11.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.23
Lazify==0.4.0
marshmallow==3.21.1
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.13.3
orjson==3.9.15
packaging==23.2
pandas==2.2.1
protobuf==4.25.3
pyarrow==15.0.1
pyarrow-hotfix==0.6
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.11.1
pytz==2024.1
PyYAML==6.0.1
ragas==0.1.3
regex==2023.12.25
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.27.0
syncer==2.0.3
tenacity==8.2.3
tiktoken==0.6.0
tomli==2.0.1
tqdm==4.66.2
types-requests==2.31.0.20240311
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
uptrace==1.22.0
urllib3==2.2.1
uvicorn==0.23.2
watchfiles==0.20.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
aiofiles==23.2.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asyncer==0.0.2
chainlit==0.7.700
filelock==3.13.1
filetype==1.2.0
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
